+++ src/store/index.ts
@@ -1,0 +1,3 @@
+export * from "./memos";
+export * from "./setting";
+export * from "./themes";
