+++ src/store/memos.ts
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+
+
+export const memoListAtom = atom<{ content: string; date: string; theme: string; time: string }[]>([]);
